Preliminary unstable push for jop40d. The current build status is listed below.
Working or semi-working:
* Wi-Fi works, but with the standard jzo54k kernel prebuilt has stability issues, and requires Wi-Fi to be disabled and reenabled. An updated kernel prebuilt will be provided in a future commit to resolve this.

* Cellular data works, but the signal strength indications and data are unavailable (to be added).

* Quick settings should work if the user setup process is taken to completion, otherwise it may not appear (to be resolved).

Not working or not configured/present:
* Wi-Fi tethering does not work. The kernel prebuilt mentioned above has not resolved the issue.

* Bluetooth (to be added).

Software or hardware features not mentioned have not been tested and are of unknown status.

End users can obtain the Wi-Fi kernel fix by cloning the kernel/tegra tree, using kernel/common as a remote, and then git-pulling from common into tegra, accepting the remote for conflicts.